feat: migrate from Netlify CMS to Decap CMS and update dependencies (#79)
* Replace Netlify CMS with Decap CMS in configuration and dependencies

* updated dependencies to latest version and corrected errors

# Conflicts:
#	package-lock.json
#	package.json

* Fix Gatsby config module error by renaming to .cjs

* fix minor ci  errors

* changed gatsby-config.cjs to gatsby-config.js to fix ci checks

* errors fixes

* Rename eslint.config.js to eslint.config.mjs

* Update gatsby-node.js

* Update gatsby-node.js

* Update gatsby-node.js

* Update gatsby-node.js to handle checks

* handled createresolvers

* parsing error updated

* resolved graphql error and other fixes

* fixed eslint 4 errors

* Resolved merge conflicts in package.json and updated dependencies

* Resolved merge conflicts in package.json and package-lock.json

* resolved final issues

* fixing jenkins ci checks more

* feat: add `js-search` (#143)

* feat: add js-search

* fix: minore change

* chore: update search list

* chore: support dark mode

* chore: add newline

* chore: made found text bold

* chore: remove titles from place

* chore: resolve file path and add commit

* Added light mode map image for landing page (#145)


* feat: enhance "Tell Your Story" section (#118)

* Added enhanced UserStoryCard to display all stories

* Added Stats Bar to Map Page

added stats bar with total stories, locations and industries

* Update src/components/StatsCard.module.css


* updated user story pop in map page

* centered read more button

* updated map screenshot in index.jsx

- updated screenshot
- linked screenshot to map page
- other fixes

* Update MapPage.css

* enhanced "tell your story" in all.jsx

- improved the UI of the tell your story section

* Update all.css

- made the theme same as with the utility class 'bg-dark'

* fixed light mode issue

* Update src/pages/all.css


* fixed pseudo class styling

* added modal and user_story PR template

* fixed PR url

* added template to PR link

---------


* chore: add support for light mode (#147)

* chore: add support for light mode andd maje badge responsive

* chore: improve stories title for ligth

* chore: revert Ui to original

---------


* fix: fix "Share Your Story" feature (#152)

* Added enhanced UserStoryCard to display all stories

* Added Stats Bar to Map Page

added stats bar with total stories, locations and industries

* Update src/components/StatsCard.module.css


* updated user story pop in map page

* centered read more button

* updated map screenshot in index.jsx

- updated screenshot
- linked screenshot to map page
- other fixes

* Update MapPage.css

* enhanced "tell your story" in all.jsx

- improved the UI of the tell your story section

* Update all.css

- made the theme same as with the utility class 'bg-dark'

* fixed light mode issue

* Update src/pages/all.css


* fixed pseudo class styling

* added modal and user_story PR template

* fixed PR url

* added template to PR link

* fixed success story submission workflow

* Update src/pages/all.jsx


* removed user-story PR template

* Update src/pages/all.jsx


* Update src/pages/all.jsx


---------


* feat: add scroll-triggered animations (#140)

* Add Scroll-triggered Animations

* Update src/pages/index.css


* Update src/pages/index.css


* Update index.css

* Update index.css

---------


* fix: update README.md to include formating command (#138)

* readme changes and format changes

* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* run format once again

* Update README.md


* Update README.md


* Update README.md


* Update README.md


* format on the new jssearch pr

* Update README.md


* Update README.md


* Update SearchContainer.css


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update index.css


---------


* feat: enhance storycard ui (#149)

* card ui

* style: update card UI and layout adjustments

* run format

* light mode

* Update src/components/StatsCard.module.css


* Update src/components/StatsCard.module.css


* Update src/components/UserStoryCard.module.css


* Update src/components/UserStoryCard.module.css


* Update src/pages/index.css


* Update src/pages/index.css


* Update src/pages/index.css


* Update src/pages/index.jsx


* Update src/pages/index.jsx


* Update src/pages/index.css


* Update index.css


* Update index.css


---------


* modified

* few changes

* trying to fix gatsby issue

* Update .gitignore

* few more additions

* fixes

* Trigger build

* fix: update Jenkinsfile to add config for build environment

* fix: update variable value in Jenkinsfile

* chore: update repo config

* feat: comment out lint and test step in Jenkinsfile

* feat: uncomment and update lint and test step in Jenkinsfile

* fix: update Jenkinsfile

* feat: update package.json and graphql queries

* fix: debug Jenkinsfile

* feat: apply suggestions from code review

* wip(Jenkinsfile) remove `NETLIFY` environment variable to avoid gatsby builidng to `/opt`

- Ref. https://github.com/gatsbyjs/gatsby/blob/cbb7ff8a39b9e2ff7e3a22d0983c3976ae6700a6/packages/gatsby-adapter-netlify/src/index.ts#L25-L29
- No consumption of the variable in the code: https://github.com/search?q=repo%3Abiru-codeastromer%2Fstories%20netlify&type=code

---------
