--- .eslintignore
@@ -1,11 +1,0 @@
-node_modules/
-.cache/
-public
-
-src/layout.css
-src/layout.jsx
-
-src/images/user-story/
-src/user-story/
-src/pages/images/
-
