+++ eslint.config.mjs
@@ -1,0 +1,98 @@
+import importPlugin from 'eslint-plugin-import';
+import reactPlugin from 'eslint-plugin-react';
+import promisePlugin from 'eslint-plugin-promise';
+import babelParser from '@babel/eslint-parser';
+
+export default [
+    {
+        ignores: [
+            'node_modules/',
+            '.cache/',
+            'public/',
+            'src/layout.css',
+            'src/layout.jsx',
+            'src/images/',
+            'src/user-story/',
+            'src/cms/',
+        ],
+    },
+    {
+        languageOptions: {
+            parser: babelParser,
+            parserOptions: {
+                ecmaVersion: 'latest',
+                jsx: true,
+                requireConfigFile: false,
+                babelOptions: {
+                    presets: ['@babel/preset-react'],
+                },
+                sourceType: 'module',
+            },
+            globals: {
+                __PATH_PREFIX__: true,
+                __dirname: true,
+                module: true,
+                require: true,
+                exports: true,
+                console: true,
+            },
+        },
+        plugins: {
+            import: importPlugin,
+            react: reactPlugin,
+            promise: promisePlugin,
+        },
+        rules: {
+            'comma-spacing': 2,
+            'eol-last': 2,
+            'import/extensions': ['error', 'ignorePackages', {js: 'never', jsx: 'never'}],
+            'import/no-unresolved': [2, {ignore: ['@gatsbyjs/reach-router']}],
+            'indent': ['error', 4],
+            'jsx-quotes': 2,
+            'key-spacing': [2],
+            'max-len': 0,
+            'no-console': 0,
+            'no-extra-semi': 2,
+            'no-multi-spaces': 'error',
+            'no-trailing-spaces': [2, {skipBlankLines: true}],
+            'no-undef': 2,
+            'no-underscore-dangle': [0],
+            'no-unused-vars': [2],
+            'no-var': 2,
+            'object-curly-spacing': ['error', 'never'],
+            'object-shorthand': [0, 'always'],
+            'prefer-arrow-callback': 2,
+            'prefer-const': 2,
+            'prefer-template': 2,
+            'quote-props': [0, 'as-needed'],
+            'quotes': [2, 'single'],
+            'react/display-name': 0,
+            'react/jsx-boolean-value': 2,
+            'react/jsx-curly-spacing': [2, {allowMultiline: true, when: 'never'}],
+            'react/jsx-equals-spacing': [2, 'never'],
+            'react/jsx-filename-extension': [1, {extensions: ['.js', '.jsx']}],
+            'react/jsx-indent': ['error', 4, {checkAttributes: true, indentLogicalExpressions: true}],
+            'react/jsx-no-undef': 2,
+            'react/jsx-one-expression-per-line': ['error', {allow: 'single-child'}],
+            'react/jsx-sort-props': 0,
+            'react/jsx-uses-react': 2,
+            'react/jsx-uses-vars': 2,
+            'react/jsx-wrap-multilines': 2,
+            'react/no-did-mount-set-state': 2,
+            'react/no-did-update-set-state': 2,
+            'react/no-multi-comp': 0,
+            'react/no-unknown-property': 2,
+            'react/prop-types': 2,
+            'react/react-in-jsx-scope': 2,
+            'react/self-closing-comp': 2,
+            'semi': [2],
+            'space-before-function-paren': [2, {anonymous: 'always', named: 'never'}],
+            'strict': [2, 'global'],
+        },
+        settings: {
+            'import/extensions': ['.js', '.jsx'],
+            'import/resolver': {node: {extensions: ['.js', '.jsx']}},
+            'react': {version: 'detect'},
+        },
+    },
+];
