+++ gatsby-config.js
@@ -1,4 +1,9 @@
 module.exports = {
+    flags: {
+        PRESERVE_WEBPACK_CACHE: true,
+        FAST_DEV: true,
+        DEV_SSR: false
+    },
     siteMetadata: {
         siteUrl: 'https://stories.jenkins.io',
         description: 'Jenkins is the way',
@@ -18,10 +23,10 @@ module.exports = {
             },
         },
         {
-            resolve: 'gatsby-plugin-netlify-cms',
+            resolve: 'gatsby-plugin-decap-cms',
             options: {
-                modulePath: require.resolve('./src/cms/netlify.jsx')
-            }
+                modulePath: `${__dirname}/src/cms/cms.js`, // Path to your CMS configuration
+            },
         },
         {
             resolve: 'gatsby-plugin-matomo',
@@ -31,7 +36,6 @@ module.exports = {
                 siteUrl: 'https://stories.jenkins.io'
             }
         },
-        'gatsby-plugin-react-helmet',
         'gatsby-plugin-sitemap',
         'gatsby-plugin-image',
         'gatsby-plugin-sharp',
@@ -82,6 +86,5 @@ module.exports = {
                 dest: `${__dirname}/schema.graphql`,
             },
         },
-        'gatsby-plugin-netlify'
     ]
 };
