+++ gatsby-node.js
@@ -4,22 +4,22 @@ const path = require('path');
 async function createUserStoryPages({graphql, createPage, createRedirect}) {
     const userStory = path.resolve('src/pages/_user_story.jsx');
     const result = await graphql(`{
-      stories: allUserStory {
-        edges {
-          node {
-            id
-            slug
-          }
-          next {
-            title
-            slug
-          }
-          previous {
-            title
-            slug
-          }
+        stories: allUserStory {
+            edges {
+                node {
+                    id
+                    slug
+                }
+                next {
+                    title
+                    slug
+                }
+                previous {
+                    title
+                    slug
+                }
+            }
         }
-      }
     }`);
 
     if (result.errors) {
@@ -29,7 +29,6 @@ async function createUserStoryPages({graphql, createPage, createRedirect}) {
 
     result.data.stories.edges.forEach(edge => {
         if (!edge.node.slug.startsWith('jenkins-is-the-way-')) {
-            // just in case handle any urls that previously had jenkins-is-the-way in the url
             createRedirect({
                 fromPath: `/user-story/jenkins-is-the-way-${edge.node.slug}/`,
                 toPath: `/user-story/${edge.node.slug}/`,
@@ -43,7 +42,7 @@ async function createUserStoryPages({graphql, createPage, createRedirect}) {
                 id: edge.node.id,
                 next: edge.next,
                 previous: edge.previous,
-            }
+            },
         });
     });
 }
@@ -70,9 +69,11 @@ exports.onCreateNode = async ({node, actions, loadNodeContent, createNodeId, cre
                     type: 'UserStory',
                 },
             };
-            const paragraphs = obj.body_content.paragraphs;
 
-            yamlNode.body_content.paragraphs = paragraphs.map((_, idx) => createNodeId(`${yamlNode.id} >>> ${idx} >>> MarkdownRemark`));
+            const paragraphs = obj.body_content.paragraphs;
+            yamlNode.body_content.paragraphs = paragraphs.map((_, idx) =>
+                createNodeId(`${yamlNode.id} >>> ${idx} >>> MarkdownRemark`)
+            );
             yamlNode.internal.contentDigest = createContentDigest(yamlNode);
 
             createNode(yamlNode);
@@ -101,31 +102,33 @@ exports.onCreateNode = async ({node, actions, loadNodeContent, createNodeId, cre
     }
 };
 
-exports.createSchemaCustomization = ({actions: {createTypes}}) => {
+exports.createSchemaCustomization = ({actions}) => {
+    const {createTypes} = actions;
     createTypes(`
+        scalar CustomDate
+
         type UserStoryMetadata {
-          build_tools: [String]
-          community_supports: [String]
-          company: String
-          company_website: String
-          industries: [String]
-          organization: String
-          platforms: [String]
-          plugins: [String]
-          programming_languages: [String]
-          project_funding: String
-          project_website: String
-          summary: String
-          team_members: [String]
-          version_control_systems: [String]
+            build_tools: [String]
+            community_supports: [String]
+            company: String
+            company_website: String
+            industries: [String]
+            organization: String
+            platforms: [String]
+            plugins: [String]
+            programming_languages: [String]
+            project_funding: String
+            project_website: String
+            summary: String
+            team_members: [String]
+            version_control_systems: [String]
         }
 
         type UserStoryBody_content @dontinfer {
-          title: String
-          paragraphs: [MarkdownRemark] @link
+            title: String
+            paragraphs: [MarkdownRemark] @link
         }
-
-  `);
+    `);
 };
 
 exports.onCreateWebpackConfig = ({stage, loaders, actions}) => {
