+++ package.json
@@ -14,42 +14,59 @@
     "serve": "gatsby serve",
     "clean": "gatsby clean",
     "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
     "format": "prettier --write \"src/**/*.{css,js,jsx,json}\"",
     "format-check": "prettier --check \"src/**/*.{css,js,jsx,json}\""
   },
   "dependencies": {
-    "cheerio": "^1.0.0-rc.10",
-    "eslint": "^8.12.0",
-    "eslint-config-react-app": "^7.0.0",
-    "eslint-plugin-promise": "^7.0.0",
-    "gatsby": "^4.10.3",
-    "gatsby-plugin-extract-schema": "^0.2.1",
-    "gatsby-plugin-image": "^2.10.1",
+    "cheerio": "^1.0.0",
+    "decap-cms": "^3.4.0",
+    "decap-cms-app": "^3.4.0",
+    "eslint-config-react-app": "^7.0.1",
+    "gatsby": "^5.14.0",
+    "gatsby-plugin-decap-cms": "^4.0.4",
+    "gatsby-plugin-extract-schema": "^0.2.2",
+    "gatsby-plugin-image": "^3.14.0",
     "gatsby-plugin-matomo": "^0.17.0",
-    "gatsby-plugin-netlify": "^5.0.0",
-    "gatsby-plugin-netlify-cms": "^6.17.0",
-    "gatsby-plugin-react-helmet": "^5.10.0",
-    "gatsby-plugin-sharp": "^4.10.2",
-    "gatsby-plugin-sitemap": "^5.10.2",
-    "gatsby-remark-images": "^6.10.2",
-    "gatsby-source-filesystem": "^4.10.1",
-    "gatsby-transformer-remark": "^5.11.1",
-    "gatsby-transformer-sharp": "^4.10.0",
+    "gatsby-plugin-sharp": "^5.14.0",
+    "gatsby-plugin-sitemap": "^6.14.0",
+    "gatsby-remark-images": "^7.14.0",
+    "gatsby-source-filesystem": "^5.14.0",
+    "gatsby-transformer-remark": "^6.14.0",
+    "gatsby-transformer-sharp": "^5.14.0",
+    "graphql": "16.10.0",
+    "hash-wasm": "^4.12.0",
     "js-search": "^2.0.1",
     "leaflet": "^1.7.1",
-    "netlify-cms-app": "^2.15.72",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
     "react-immutable-proptypes": "^2.2.0",
-    "react-leaflet": "^3.2.5",
-    "remark-html": "^16.0.0",
+    "react-leaflet": "^4.2.1",
+    "remark": "^15.0.1",
+    "remark-html": "^16.0.1",
     "truncate": "^3.0.0",
-    "yaml": "^2.0.0"
+    "yaml": "^2.0.1"
   },
   "devDependencies": {
+    "@babel/eslint-parser": "^7.25.9",
+    "@jenkinsci/gatsby-plugin-jenkins-layout": "^1.5.0",
+    "eslint": "^8.57.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-react": "^7.37.2"
+  },
+  "resolutions": {
+    "graphql": "16.10.0"
+  },
+  "overrides": {
+    "graphql": "16.10.0",
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@jenkinsci/gatsby-plugin-jenkins-layout": "^1.5.0",
     "prettier": "^3.4.2"
+  },
+  "optionalDependencies": {
+    "@parcel/watcher": "^2.5.1"
   }
 }
